<Enhancement>[]: <Inter/Intra Quadrant Move works>
[
* Star Chart works
* Long range sensors work
* Phasers fire but amount not controllable
* Photon torpedoes first at all targets
]

[https://github.com/hasii2011/pytrek/issues/115]

+++ pytrek/commandparser/ManualMoveData.py
@@ -5,5 +5,5 @@ from dataclasses import dataclass
 @dataclass
 class ManualMoveData:
 
-    deltaX: int = 0
-    deltaY: int = 0
+    deltaX: float = 0.0
+    deltaY: float = 0.0
